requirements.txt
bcrypt==4.1.2
argon2-cffi==23.1.0
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
